+++ Editor/ActiveScene_PropertyDrawer.cs
@@ -17,7 +17,6 @@
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.SceneManagement;
-using UnityEditor.SceneManagement;
 using HH.MultiSceneToolsEditor;
 
 // View